lab07/requirements.txt
agent-framework --pre
azure-identity
python-dotenv
